removed old exceptions without python, Exceptions.* now can only be run under python environment
